Split validateEntry's typed checks into config-validation-core
